requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
databases[aiosqlite]==0.8.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
RPi.GPIO==0.7.1
python-dotenv==1.0.0
asyncio-mqtt==0.13.0
loguru==0.7.2
